fix: review round on the discoverability PR + the test import cycle
The registry round-trip tests (whitelist vs live resource registry) move
to the external subscriptions_test package: internal/resources now
imports internal/subscriptions for the recorder's marker, so an internal
test importing resources back was an import cycle CI's repo-wide lint
caught and my package-scoped local lint missed.

CodeRabbit round: the icon-contract sentence uses the wire-cased field
names (sizes/theme); the marker parity test is table-driven per template
with an expected-result name; gen_llms's subscription summaries no
longer say all 26 kinds accept resources/subscribe (they name
subscriptions/listen first with the stateful condition) nor call the
three single-parent lists single-object kinds. llms catalogs
regenerated.
